Refactored the routes into controllers but broke all the include_chain_clauses_in_custom_matcher_descriptions
