+++ error.html
@@ -1,0 +1,20 @@
+<html>
+<head>
+  <style type="text/css">
+  body { text-align:center;font-family:helvetica,arial;font-size:22px;
+    color:#888;margin:20px}
+  #c {margin:0 auto;width:500px;text-align:left}
+  </style>
+</head>
+<body>
+  <h2>Sinatra doesn’t know this ditty.</h2>
+  <img src='http://example.org/__sinatra__/404.png'>
+  <div id="c">
+    Try this:
+    <pre>get &#x27;&#x2F;radius&#x2F;80238&#x2F;999&#x27; do
+  &quot;Hello World&quot;
+end
+</pre>
+  </div>
+</body>
+</html>
