+++ spec/models/bad_request_spec.rb
@@ -1,5 +1,4 @@
 require "spec_helper"
-require './zipcode_base_api'
 require './models/bad_request'
 
 RSpec.describe BadRequest do
